fix: fix several small issues
Fixed timestamp formatting to correctly calculate timestamp.
The 'calculation' used string concatenation, not addition.

Removed broken test cases that called non-existing methods.

Renamed classes to start with capital letter.
